Match device name on Linux
On Linux, the Model:Cycles was not detected.

This happens because the device name on Linux is reported with an additional
suffix, i.e. "Elektron Model:Cycles MIDI 1", but the test was looking for an
exact match with "Elektron Model:Cycles"

This changes the test to match only the start of the string, ignoring
any additional text after the device name.